refactored, reorganized python code
