+++ python/pygfdm/modulation.py
@@ -1,10 +1,25 @@
-#!/usr/bin/env python2.7
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+# Copyright 2016 Andrej Rode.
+#
+# This is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 3, or (at your option)
+# any later version.
+#
+# This software is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this software; see the file COPYING.  If not, write to
+# the Free Software Foundation, Inc., 51 Franklin Street,
+# Boston, MA 02110-1301, USA.
+#
 
 import numpy as np
-import commpy as cp
-import matplotlib.pyplot as plt
-# from synchronization import *
-from gfdm_plot_utils import plot_gfdm_matrix
 from filters import gfdm_filter_taps, gfdm_freq_taps, gfdm_freq_taps_sparse
 from mapping import reshape_input
