+++ python/pygfdm/utils.py
@@ -1,72 +1,61 @@
-#!/usr/bin/env python2.7
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+# Copyright 2016 Andrej Rode, Johannes Demel.
+#
+# This is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 3, or (at your option)
+# any later version.
+#
+# This software is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this software; see the file COPYING.  If not, write to
+# the Free Software Foundation, Inc., 51 Franklin Street,
+# Boston, MA 02110-1301, USA.
+#
 
-import matplotlib.pyplot as mp
-from mpl_toolkits.mplot3d import Axes3D
-from matplotlib import cm
+
 import numpy as np
 
-__all__ = ['plotFFT', 'plotTransmitMatrix']
 
-def plotFFT(x):
-    fig = mp.figure()
-    ax = fig.add_subplot(111)
-    x_fft = np.fft.fft(x, n=x.shape[-1])
-    x_t = np.fft.fftfreq(x.shape[-1])
-    ax.stem(x_t, np.abs(x_fft))
-    fig.show()
+def get_random_qpsk(nsamples):
+    d = np.random.randint(0, 2, 2 * nsamples) * -2. + 1.
+    d = np.reshape(d, (2, -1))
+    d = d.astype(dtype=np.complex)
+    return d[0] + 1j * d[1]
 
 
-def plotTransmitMatrix(A):
-    fig = mp.figure()
-    ax = fig.add_subplot(111, projection='3d')
-    x, y = np.meshgrid(np.arange(A.shape[-1]), np.arange(A.shape[-2]))
-    ax.plot_wireframe(x, y, np.abs(A),cstride=1,rstride=1)
-    #ax.plot_surface(x, y, np.abs(A), rstride=1, cstride=1, cmap=cm.Greys, linewidth=0, antialiased=False)
-    fig.show()
+def get_random_samples(nsamples):
+    d = np.random.standard_normal(2 * nsamples)
+    d = np.reshape(d, (2, -1))
+    return d[0] + 1j * d[1]
 
 
-def plotSymbols(x):
-    fig = mp.figure()
-    ax = fig.add_subplot(211)
-    ax.plot(x.real)
-    ax2 = fig.add_subplot(212)
-    ax2.plot(x.imag)
-    fig.show()
-
-def compareRx(*args):
-    fig = mp.figure()
-    ax = fig.add_subplot(211)
-    for x in args:
-        ax.plot(x.real)
-    ax = fig.add_subplot(212)
-    for x in args:
-        ax.plot(x.imag)
-    fig.show()
-
-def plotScatter(ax,x):
-    ax.scatter(x.real,x.imag)
-
-
-def plotBER(x):
+def randomQAMSymbols(length, M):
     '''
-    nested dict - structure (see testsuite.py)
+     length: number of symbols to generate
+     M: M-QAM - Order (4,16,64,...)
     '''
-    fig = mp.figure()
-    ax = fig.add_subplot(111)
-    for m in x:
-        for k in x[m]:
-            for qam in x[m][k]:
-                ax.plot(x[m][k][qam][0],x[m][k][qam][1], label="M:{},K:{},QAM:{}".format(m,k,qam))
-    ax.legend(bbox_to_anchor=(0.95, 1), loc=2, borderaxespad=0.)
-    fig.show()
+    n = np.sqrt(M / 4)
+    if np.around(n) - n > 1e-10:
+        raise Exception('M must be power of 4')
+    n = int(n)
+    n_M_pos = np.array([1 + 2 * i for i in xrange(n)])
+    n_M_neg = np.array([-1 - 2 * i for i in xrange(n)])
+    choices = np.concatenate((n_M_pos, n_M_neg))
+    return np.array(
+        [np.random.choice(choices) + 1j * np.random.choice
+        (choices) for i in xrange(length)])
 
 
-def plot_sc_int(x,k_range,m_range,qam_range,j):
-    fig = mp.figure()
-    ax = fig.add_subplot(111)
-    for k in k_range:
-        for m in m_range:
-            for qam in qam_range:
-                ax.plot([x[k][m][qam][j_it] for j_it in xrange(j)], label="K:{}, M:{},\n QAM:{}".format(k,m,qam))
-    ax.legend(bbox_to_anchor=(0.95,1), loc=2, borderaxespad=0.)
-    fig.show()
+def get_zero_f_data(k, K, M):
+    data = np.zeros(K)
+    data[k] = 1.
+    # data = np.tile(data, M)
+    data = np.repeat(data, M)
+    return data
