+++ python/qa_modulator_cc.py
@@ -23,8 +23,9 @@ from gnuradio import gr, gr_unittest
 from gnuradio import blocks
 import gfdm_swig as gfdms
 from pygfdm.gfdm_modulation import gfdm_modulate_block
-from pygfdm.mapping import get_data_matrix, get_random_qpsk
+from pygfdm.mapping import get_data_matrix
 from pygfdm.filters import get_frequency_domain_filter
+from pygfdm.utils import get_random_qpsk
 import numpy as np
